requirements.txt
rich = "13.7.1"
textual = "0.56.4"
python-dotenv = "1.0.1"